Release build: 10.4.1
Release Change Notes:
Enhancements

- Major release with support for Enhanced Replay. This will involve several changes with the TealeafLayoutConfig.json file which now has a 'IBMGlobalScreenSettings'. IBMGlobalScreenSettings represents a global settings for all the other screens. Several properties have changed and have been deprecated.

ScreenChange
This used to be 'do', which is now obsolete. Please use ScreenChange key from Tealeaf SDK 10.4.1 onwards.

DisplayName
This used to be 'screenViewName', which is now obsolete. Please use DisplayName key from Tealeaf SDK 10.4.1 onwards. DisplayName is the name that will be displayed during replay in the navigation list. If DisplayName is empty, view controller class name will be used.

CaptureLayoutDelay
This used to be called 'delay', which is now obsolete. Please use CaptureLayoutDelay key from Tealeaf SDK 10.4.1 onwards. This is delay in milliseconds before Tealeaf SDK takes layout capture of a screen.

ScreenShot
This used to be called 'takeScreenShot', which is now obsolete. Please use ScreenShot key from Tealeaf SDK 10.4.1 onwards.  However before it applied only for type 2 load events. Now this applies to type 2, 4, 10, 11. If you do not want Tealeaf SDK to take any screen shots on a given screen, set ScreenShot too false.

NumberOfWebViews
This used to be 'numberOfWebviews'. Now its a combination of 'numberOfWebviews' and 'isWebView'. Setting NumberOfWebViews = 0 is as good as 'isWebView' = false. Non zero value NumberOfWebViews indicates number of web views on a view controller. 'numberOfWebviews' and 'isWebView' are now obsolete. Please use NumberOfWebViews.

CaptureUserEvents
This used to be 'CaptureScreenContents', which is now obsolete. Please use CaptureUserEvents key from Tealeaf SDK 10.4.1 onwards.  If CaptureUserEvents is set to 1, Tealeaf SDK will capture user events (type 2,11) on a given screen. If set to 0, Tealeaf SDK will not capture type 2,11 but based on the value of CaptureScreenVisits will capture screen load/unload events.

CaptureScreenVisits
If you do not want to track screen load unload events, set this to false. CaptureScreenVisits applies only if CaptureUserEvents = false. CaptureUserEvents is true CaptureScreenVisits is ignored and load/unload are tracked; because it doesn't make any sense to track user events on a screen whose visits are not tracked.

CaptureLayoutOn
There are 3 possible events on which Tealeaf SDK can capture layouts. If CaptureLayoutOn is set to 2, Tealeaf SDK will capture the layout as soon as view controller is loaded. If CaptureLayoutOn is set to 1, Tealeaf SDK will capture the layout after end user makes a first gesture on a given view controller. If CaptureLayoutOn is set to 0, layout will not be captured.

CaptureScreenshotOn
There are 3 possible events on which Tealeaf SDK can capture screen shots for view controller load event. If CaptureScreenshotOn is set to 2, Tealeaf SDK will capture the screen shot as soon as view controller is loaded. If CaptureScreenshotOn is set to 1, Tealeaf SDK will capture the screen shot after end user makes a first gesture on a given view controller. If CaptureScreenshotOn is set to 0, screen shot will not be captured. Note that even if CaptureScreenshotOn is set to 0; if ScreenShot is true, Tealeaf SDK will continue to capture screen shots on other user events such as type 2 and type 11. CaptureScreenshotOn applies only to screen shots on view controller load.
The following snippet shows an example of the TealeafLayoutConfig.json file.

{
   'AutoLayout': {
       'IBMGlobalScreenSettings':{
           'ScreenChange': false,
           'DisplayName': '',
           'CaptureLayoutDelay': 0,
           'ScreenShot': true,
           'NumberOfWebViews': 0,
           'CaptureUserEvents': true,
           'CaptureScreenVisits': true,
           'CaptureLayoutOn': 2,
           'CaptureScreenshotOn': 0
       },
       'PaymentViewController':{
           'ScreenChange': false,
           'DisplayName': 'The Payment Screen',
           'CaptureLayoutDelay': 0,
           'ScreenShot': false,
           'NumberOfWebViews': 0,
           'CaptureUserEvents': false,
           'CaptureScreenVisits': false,
           'CaptureLayoutOn': 2,
           'CaptureScreenshotOn': 0
       }
   },
   'AppendMapIds': {
       '[w,9290],[v,0]': {
           'mid': 'ASimpleUIView'
       },
       'tag2999999': {
           'mid': 'giveAdditionalId1'
       },
       'idxPathValue': {
           'mid': 'giveAdditionalId2'
       }
   }
}

By default, if you are missing the IBMGlobalScreenSettings section, you have the following defaults, which capture all pages, events, and gestures but no screenshots:

'AutoLayout': {
'IBMGlobalScreenSettings':{
           'ScreenChange': false,
           'DisplayName': '',
           'CaptureLayoutDelay': 0,
           'ScreenShot': false,
           'NumberOfWebViews': 0,
           'CaptureUserEvents': true,
           'CaptureScreenVisits': true,
           'CaptureLayoutOn': 2,
           'CaptureScreenshotOn': 0
       },
},
'AppendMapIds': {
   }

For more information review: https://developer.ibm.com/customer-engagement/tutorials/configuring-watson-customer-experience-analytics-mobile-basic-edition-sdk-ios-development-app/

- Changed relxy from 2 to 4 decimal points for Overstat precision.
- Adding additional Enhanced Replay functionality into library.
- Due to adding IBMGlobalScreenSettings section in TealeafLayoutConfig.json with a delay property, we are removing AutoWebViewCaptureDelay and DefaultAutoLayoutDelay from TealeafAdvancedConfig.json which is no longer be referenced. So please adjust your IBMGlobalScreenSettings section in TealeafLayoutConfig.json.

Bug Fixes

- None.

Known Issues

- Actionsheets with webviews are not fully supported replay gets broken up into different pages.
- Date Picker change events are no longer replayed on SaaS and On-Prem for regular replay. These work correctly with Enhanced Replay.
- Hybrid session sometime load javascript bridge late and it is not possible to get dcid correctly. So you will have an error 'dcid=dcid-none. The page was not captured correctly' for the first page. You you selected next page. It should be fixed afterwards.
- Captures click event twice on all table view controller.
- Support for arm64e will be added soon, Apple has not completed the specification. Currently it is just beta.
- React Native has an issue for dynamic builds. We will use static version instead.
- Alerts are broken on replay. It will be fixed in next version.
- On iOS 12.x SDK will not capture carrier name. It will be recorded 'Carrier'. This is because 'Apple API [[CTTelephonyNetworkInfo alloc] init]' on iOS 12 is not stable.

The following issues are for Enhanced Replay:
- When configured to capture nothing for a page, the library captures the UIAlertController screen.